move think object to components subdir
