+++ src/ecs/components/think.c
