+++ src/ecs/components/think.h
